Bug 1702410 - Rename backfill bot' s packages to Sherlock
