+++ tests/perfalert/sherlock/test_backfill_reports/test_backfill_report_maintainer.py
@@ -1,7 +1,7 @@
 import datetime
 from typing import Tuple
 
-from treeherder.perf.auto_perf_sheriffing.backfill_reports import (
+from treeherder.perf.sherlock.backfill_reports import (
     BackfillReportMaintainer,
 )
 from treeherder.perf.models import (
