+++ tests/perfalert/sherlock/test_backfill_tool.py
@@ -1,6 +1,6 @@
 import pytest
 
-from treeherder.perf.auto_perf_sheriffing.backfill_tool import BackfillTool
+from treeherder.perf.sherlock.backfill_tool import BackfillTool
 from treeherder.perf.exceptions import CannotBackfill
 from treeherder.services.taskcluster import TaskclusterModelProxy
