+++ treeherder/perf/sherlock/factories.py
@@ -2,14 +2,14 @@ from datetime import timedelta
 
 from django.conf import settings
 
-from treeherder.perf.auto_perf_sheriffing.backfill_reports import (
+from treeherder.perf.sherlock.backfill_reports import (
     BackfillReportMaintainer,
     AlertsPicker,
     IdentifyAlertRetriggerables,
 )
-from treeherder.perf.auto_perf_sheriffing.backfill_tool import BackfillTool
-from treeherder.perf.auto_perf_sheriffing.perf_sheriff_bot import PerfSheriffBot
-from treeherder.perf.auto_perf_sheriffing.secretary_tool import SecretaryTool
+from treeherder.perf.sherlock.backfill_tool import BackfillTool
+from treeherder.perf.sherlock.perf_sheriff_bot import PerfSheriffBot
+from treeherder.perf.sherlock.secretary_tool import SecretaryTool
 from treeherder.services.taskcluster import DEFAULT_ROOT_URL, notify_client_factory
 from treeherder.services.taskcluster import TaskclusterModelProxy
